MM-62079: Attempt#n to fix TestCreatePost
The issue is somehow related to Redis
as I see this always happens on Postgres CI jobs.

The problem is with the roles. This block:

```
defaultPerms := th.SaveDefaultRolePermissions()
defer th.RestoreDefaultRolePermissions(defaultPerms)

th.RemovePermissionFromRole(model.PermissionUseChannelMentions.Id, model.ChannelUserRoleId)
```

somehow either does not remove the permission, or even
if it does, while trying to get the permission
we get a stale entry.

I have been trying to make it to reproduce in all possible
ways but unable to.

This PR is just an effort to see if MaxFlushDelay is a potential
suspect or not. If this also does not work, I think
we have to go back to skipping this test again.

https://mattermost.atlassian.net/browse/MM-62079

```release-note
NONE
```